+++ taiga/projects/notifications/services.py
@@ -31,14 +31,17 @@ from django.utils.translation import ugettext as _
 
 from taiga.base import exceptions as exc
 from taiga.base.mails import InlineCSSTemplateMail
+from taiga.front.templatetags.functions import resolve as resolve_front_url
 from taiga.projects.notifications.choices import NotifyLevel
 from taiga.projects.history.choices import HistoryType
 from taiga.projects.history.services import (make_key_from_model_object,
                                              get_last_snapshot_for_key,
                                              get_model_from_key)
 from taiga.permissions.services import user_has_perm
+from taiga.events import events
 
 from .models import HistoryChangeNotification, Watched
+from .squashing import squash_history_entries
 
 
 def notify_policy_exists(project, user) -> bool:
@@ -52,7 +55,8 @@ def notify_policy_exists(project, user) -> bool:
     return qs.exists()
 
 
-def create_notify_policy(project, user, level=NotifyLevel.involved):
+def create_notify_policy(project, user, level=NotifyLevel.involved,
+                         live_level=NotifyLevel.involved):
     """
     Given a project and user, create notification policy for it.
     """
@@ -60,23 +64,30 @@ def create_notify_policy(project, user, level=NotifyLevel.involved):
     try:
         return model_cls.objects.create(project=project,
                                         user=user,
-                                        notify_level=level)
+                                        notify_level=level,
+                                        live_notify_level=live_level)
     except IntegrityError as e:
-        raise exc.IntegrityError(_("Notify exists for specified user and project")) from e
+        raise exc.IntegrityError(
+            _("Notify exists for specified user and project")) from e
 
 
-def create_notify_policy_if_not_exists(project, user, level=NotifyLevel.involved):
+def create_notify_policy_if_not_exists(project, user,
+                                       level=NotifyLevel.involved,
+                                       live_level=NotifyLevel.involved):
     """
     Given a project and user, create notification policy for it.
     """
     model_cls = apps.get_model("notifications", "NotifyPolicy")
     try:
-        result = model_cls.objects.get_or_create(project=project,
-                                                 user=user,
-                                                 defaults={"notify_level": level})
+        result = model_cls.objects.get_or_create(
+            project=project,
+            user=user,
+            defaults={"notify_level": level, "live_notify_level": live_level}
+        )
         return result[0]
     except IntegrityError as e:
-        raise exc.IntegrityError(_("Notify exists for specified user and project")) from e
+        raise exc.IntegrityError(
+            _("Notify exists for specified user and project")) from e
 
 
 def analize_object_for_watchers(obj: object, comment: str, user: object):
@@ -131,7 +142,7 @@ def _filter_notificable(user):
     return user.is_active and not user.is_system
 
 
-def get_users_to_notify(obj, *, history=None, discard_users=None) -> list:
+def get_users_to_notify(obj, *, history=None, discard_users=None, live=False) -> list:
     """
     Get filtered set of users to notify for specified
     model instance and changer.
@@ -143,6 +154,8 @@ def get_users_to_notify(obj, *, history=None, discard_users=None) -> list:
 
     def _check_level(project: object, user: object, levels: tuple) -> bool:
         policy = project.cached_notify_policy_for_user(user)
+        if live:
+            return policy.live_notify_level in levels
         return policy.notify_level in levels
 
     _can_notify_hard = partial(_check_level, project,
@@ -219,7 +232,6 @@ def send_notifications(obj, *, history):
                                             owner=owner,
                                             project=obj.project,
                                             history_type=history.type))
-
     notification.updated_datetime = timezone.now()
     notification.save()
     notification.history_entries.add(history)
@@ -233,6 +245,10 @@ def send_notifications(obj, *, history):
     if settings.CHANGE_NOTIFICATIONS_MIN_INTERVAL == 0:
         send_sync_notifications(notification.id)
 
+    live_notify_users = get_users_to_notify(obj, history=history, discard_users=[notification.owner], live=True)
+    for user in live_notify_users:
+        events.emit_live_notification_for_model(obj, user, history)
+
 
 @transaction.atomic
 def send_sync_notifications(notification_id):
@@ -243,13 +259,21 @@ def send_sync_notifications(notification_id):
     """
 
     notification = HistoryChangeNotification.objects.select_for_update().get(pk=notification_id)
-    # If the last modification is too recent we ignore it
+    # If the last modification is too recent we ignore it for the time being
     now = timezone.now()
     time_diff = now - notification.updated_datetime
     if time_diff.seconds < settings.CHANGE_NOTIFICATIONS_MIN_INTERVAL:
         return
 
     history_entries = tuple(notification.history_entries.all().order_by("created_at"))
+    history_entries = list(squash_history_entries(history_entries))
+
+    # If there are no effective modifications we can delete this notification
+    # without further processing
+    if notification.history_type == HistoryType.change and not history_entries:
+        notification.delete()
+        return
+
     obj, _ = get_last_snapshot_for_key(notification.key)
     obj_class = get_model_from_key(obj.key)
 
@@ -273,6 +297,7 @@ def send_sync_notifications(notification_id):
 
     now = datetime.datetime.now()
     format_args = {
+        "unsubscribe_url": resolve_front_url('settings-mail-notifications'),
         "project_slug": notification.project.slug,
         "project_name": notification.project.name,
         "msg_id": msg_id,
@@ -285,13 +310,15 @@ def send_sync_notifications(notification_id):
         "In-Reply-To": "<{project_slug}/{msg_id}@{domain}>".format(**format_args),
         "References": "<{project_slug}/{msg_id}@{domain}>".format(**format_args),
         "List-ID": 'Taiga/{project_name} <taiga.{project_slug}@{domain}>'.format(**format_args),
-        "Thread-Index": make_ms_thread_index("<{project_slug}/{msg_id}@{domain}>".format(**format_args), now)
+        "Thread-Index": make_ms_thread_index("<{project_slug}/{msg_id}@{domain}>".format(**format_args), now),
+        "List-Unsubscribe": "<{unsubscribe_url}>".format(**format_args),
     }
 
     for user in notification.notify_users.distinct():
         context["user"] = user
         context["lang"] = user.lang or settings.LANGUAGE_CODE
         email.send(user.email, context, headers=headers)
+
 
     notification.delete()
 
@@ -404,7 +431,11 @@ def add_watcher(obj, user):
         project=obj.project)
 
     notify_policy, _ = apps.get_model("notifications", "NotifyPolicy").objects.get_or_create(
-        project=obj.project, user=user, defaults={"notify_level": NotifyLevel.involved})
+        project=obj.project,
+        user=user,
+        defaults={"notify_level": NotifyLevel.involved,
+                  "live_notify_level": NotifyLevel.involved}
+    )
 
     return watched
 
@@ -426,22 +457,25 @@ def remove_watcher(obj, user):
     qs.delete()
 
 
-def set_notify_policy_level(notify_policy, notify_level):
+def set_notify_policy_level(notify_policy, notify_level, live=False):
     """
     Set notification level for specified policy.
     """
     if notify_level not in [e.value for e in NotifyLevel]:
         raise exc.IntegrityError(_("Invalid value for notify level"))
 
-    notify_policy.notify_level = notify_level
+    if live:
+        notify_policy.live_notify_level = notify_level
+    else:
+        notify_policy.notify_level = notify_level
     notify_policy.save()
 
 
-def set_notify_policy_level_to_ignore(notify_policy):
+def set_notify_policy_level_to_ignore(notify_policy, live=False):
     """
     Set notification level for specified policy.
     """
-    set_notify_policy_level(notify_policy, NotifyLevel.none)
+    set_notify_policy_level(notify_policy, NotifyLevel.none, live=live)
 
 
 def make_ms_thread_index(msg_id, dt):